Add "has issued errors" method on a compilation
`Compilation::getSemanticDiagnostics()` caches its result and is meant
to be called on a finished compilation for which no more diagnostics
will be issued.

For use cases where user code using slang as a library wants to issue
its own diagnostics, but doesn't want to deal with an AST which failed
elaboration, provide a `hasIssuedErrors()` method to find out early
if the elaboration failed.